Version 9.08
Fixed a bug in the restoring of user setting from prior session.  The
bug overwrote the user settings with the selected layout settings.  This
happened in the Form3 Load routine when the layout selector was loaded
from the last session.  It cause a chain reaction to load those
settings.  A variable is now used to indicate form3Loading=TRUE during
the form load.  The routine downstream now does nothing during a form
load.