Version 9.111
Rewrote the print counter code so it works between operator and KIOSK
modes
to keep track of # of prints and available paper more accurately.

Added a workaround to a photoshop cs2 bug in creating GIFs where it
truncates
file names. Emailing GIFs & copy to cloud was broken due to this.

PrintProcessorThread, the KIOSK thread, was divided into subrountines to

make it more understandable.  KIOSK prints a message to the debug window
when it rejects an image due to date age.

Email thread will email both a jpg and GIF if the matching file name is
found.

If load balancing is enabled, the main panel Printer #1/Printer #2
changes to
AutoSel P1/AutoSel P2, and the checkboxes disabled, so the operator
knows
the printer selection is under program control. This was not clear
before.

If KIOSK date qualification is NOT checked, the date select and
Print-any-with-no-dates checkbox are disabled, for operator clarity.