Version 9.12
Changed the file decoration by moving the orig file name to the end so
all decoration happens as a prefix.  This is done so the file names on
incoming emails&txt msgs dont preceed the decorations and cause problems
in the javascript.
Also added a check to make sure the Machine Name in the config panel is
3 characters in length. This is important to normalize the serial # in
the text layer for any actions that modify this layer.