Version 11.01
Big addition - Filters!  A new .CSV file has been added.  This is the
filters.csv file in the software folder.  It specifies the name, action
and action set.  The file is read at startup and the list is presented
in THREE dropdown list boxes allowing up to three filters to be selected
to combine effects.   The selected filters are passed to photoshop via
the config.txt file.  The automaticmode.js file will call the selected
filters, right before saving and printing the file.