Version 12.05
Lots of changes to allow Post View window to be used as an email
station. This facilitates the Capture/Kiosk machine to act as a master
with 2 machines as slaves.  The emailed images can then be copied to
another folder, in the cloud or forth machine, etc.

-Added 2nd cloud folder so two folder/machines can receive images from
the kiosk processor.
Added a cloud folder to the Post View window so all emailed images can
be saved in the cloud
-Added "Opt-in" and "Permit images to be used" check boxes on the
Preview and Post View email windows.
-Added a folder check to the target cloud folders to make sure the
folders do exist during runtime.  Network connections might be lost
generating execptions on the copy.
On email sends in Preview/PostView, the email address is automatically
copied for the next paste operation.
-Expanded "Send Mail" button to resend email in the printed folder as
well as the Preview (capture) folder.