Version 14.01
Fixed a memory leak in Pic2Print.  The preview window showing the
cropping marks on the image was losing memory blocks. Rewrote that
routine..

Fixed a problem in the javascript with the print count passed in.  The
file name is used to spec the print count. The JS code scanned the name
multiple times, grabbing old data from prior prints.