Version 14.02
Fixed one bug. Copying 100+ files to the c:\onsite\capture folder
flooded the cacheing causing an unhandled exception on an out-of-memory
situation.  I put a semaphore around the timer code so it doesn't
reenter on incoming flood of images.
Added a total file count preceding the Total Printed count.
Incremented version to 14.02