Version 14.09
Fixed an outlook issue. Outlook keeps all inline images within a private
cache and doesn't allow these images to be saved.  So, the actual user
images has to be sent as an attachment w/o a CID or else outlook
swallows the image.  Now the user image can be saved as intended.