Version 15.03
Found the fatal bug where the program crashed out of the blue.  The PostView window was too aggressive in scanning and reading .jpgs.  The file system had not yet flushed all info out to the disk by the time pic2print tried to load the file.  1 second wait states were in serted between the detection and loading of images which fixed the problem.  Also added a TRY/CATCH to the file loading for handling the error gracefully.